Additional structure optimization.
