+++ config/webpack/prod.js
@@ -14,26 +14,7 @@ const merge = require("webpack-merge");
 
 const commonConfig = require("./common");
 const paths = require("../paths");
-
-/**
- * Helper function to ensure that the provided path ends with a / or not,
- * depending on the requirements.
- *
- * @param path {String} The path to check for.
- * @param needsSlash {Boolean} Indicated whether it is required to have a trailing slash or not.
- * @return {String} The provided value if it already satisfies the "must (not) have" condition,
- *                  or the fixed string otherwise.
- */
-function ensureEndingSlash(path, needsSlash) {
-  const hasSlash = path.endsWith("/");
-  if (hasSlash && !needsSlash) {
-    return path.slice(0, -1);
-  } else if (!hasSlash && needsSlash) {
-    return `${path}/`;
-  } else {
-    return path;
-  }
-}
+const { ensureEndingSlash } = require("./util");
 
 /**
  * The production build may or may not include the BundleAnalyzerPlugin to visualize the build
